Refactor Person to Applicant class
