+++ src/main/java/seedu/address/model/applicant/Applicant.java
