+++ src/main/java/seedu/address/model/applicant/Email.java
