+++ src/main/java/seedu/address/model/applicant/Name.java
