+++ src/test/java/seedu/address/model/applicant/ApplicantTest.java
