+++ src/test/java/seedu/address/model/applicant/EmailTest.java
